requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
django-compositekey==1.5.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-nested-routers==0.93.3
drf-spectacular==0.20.0
drf-spectacular-sidecar==2021.9.27
inflection==0.5.1
jsonschema==4.0.1
Markdown==3.3.4
PyJWT==2.1.0
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
sqlparse==0.4.2
uritemplate==3.0.1
python-dotenv==0.19.2